Move files to constants/
