+++ src/constants/coreSettings.ts
